MILAB-6145: extract templateHasWasm into pl-model-backend
Both block-tools' pack-time detection and pl-middle-layer's install/
listing-time gating had drift-prone local copies of the same tree
walker. Hoist a single implementation into pl-model-backend (next to
TemplateWasmDataV3), accepting unknown so parsed JSON can be passed
without a prior type assertion. block-tools now depends on
pl-model-backend; its existing tests import the function from the
shared package.
