+++ tools/block-tools/package.json
@@ -37,6 +37,7 @@
   "dependencies": {
     "@aws-sdk/client-s3": "catalog:",
     "@milaboratories/pl-http": "workspace:*",
+    "@milaboratories/pl-model-backend": "workspace:*",
     "@milaboratories/pl-model-common": "workspace:*",
     "@milaboratories/pl-model-middle-layer": "workspace:*",
     "@milaboratories/resolve-helper": "workspace:*",
